create CRUD rest for institution with http status

+++ src/main/java/pl/coderslab/charity/institution/InstitutionService.java
@@ -6,6 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import pl.coderslab.charity.domain.model.Institution;
 import pl.coderslab.charity.domain.repositories.InstitutionRepository;
+import pl.coderslab.charity.rest.InvalidDataException;
 
 import java.util.List;
 
@@ -21,8 +22,11 @@ public class InstitutionService {
         return institutionRepository.findAll();
     }
 
-    public Institution save(Institution institution) {
-        return institutionRepository.save(institution);
+    public Long save(InstitutionDTO institutionDTO) {
+        Institution institution = new Institution();
+        institution.setName(institutionDTO.getName());
+        institution.setDescription(institutionDTO.getDescription());
+        return institutionRepository.save(institution).getId();
     }
 
     public ResponseEntity getById(Long id) {
@@ -33,12 +37,24 @@ public class InstitutionService {
 
     public Institution deleteById(Long id) {
         Institution institution = institutionRepository.getById(id);
+        log.debug("Institution to delete: {}", institution);
+        if (institution == null) {
+            throw new InvalidDataException();
+        }
         institutionRepository.delete(institution);
         return institution;
     }
 
 
-    public Institution update(Institution institution) {
+    public Institution update(Long id, InstitutionDTO institutionDTO) {
+        Institution institution = institutionRepository.getById(id);
+        log.debug("Institution to delete: {}", institution);
+        if (institution == null) {
+            throw new InvalidDataException();
+        }
+        institution.setId(id);
+        institution.setName(institutionDTO.getName());
+        institution.setDescription(institutionDTO.getDescription());
         return institutionRepository.save(institution);
     }
 }
